Make svm kernel evaluation faster
This is still a bit too slow. In libsvm, the kernel matrix is only
evaluated once for an entire nfoldcrossvalidation round, so there are
clear advantages to that architechture. Here, we'd need to change the
clean architechture that we have now. Maybe, it's possible to do a sort
of kernel precomputation at some point in the pipeline of models

    feature selection -> normalise -> precompute kernel -> gridsearch

Maybe it just requires a cache:

    def kernel(x,y):
        try:
            return _cache[x,y]
        except KeyError:
            _cache[x,y] = _realkernel(x,y)
            return _cache[x,y]

Precomputation works like this, but with a maddeningly efficient 'cache'
(i.e., indexing into a 2-d array).