ENH Use mapreduce in jugparallel.kmeans
Previously, a very large reduction task was generated. Even though it was a
simple task, the need to load all of the information led to it taking a long
time. Although I did not observe it, I can even imagine that this would lead to
OutOfMemory errors that are prevented by using jug.mapreduce.