Added initial centroid parameter for kmeans
There are many applications where specifying the initial centroids for a
kmeans run is useful.  This could be in cases where new initialization
methods are being considered or when some iterative kmeans algorithm is
desired (ie: xmeans)

This pull request adds the base functionality for this in
milk.unsupervised.kmeans which should extend to all other related kmeans
functions.